refactor: styles
