+++ src/theme/index.ts
@@ -1,4 +1,8 @@
+import withStyles from './styles/styles';
+
 export * from './animations';
 export * from './styles';
 export * from './defaultTheme';
 export * from './theme';
+
+export { withStyles };
